requirements.txt
discord.py==2.5.2
jsonpickle==4.0.0
numpy==2.1.3
python-dotenv==1.1.0
